requirements.txt
flet==0.27.6
sentence-transformers==2.5.1
torch>=2.0.0
numpy>=1.24.0
sounddevice>=0.4.6
vosk>=0.3.44
pyaudio>=0.2.13
gtts>=2.3.2
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-dotenv==1.0.1
pydantic>=2.5.3
supabase>=2.3.4
asyncpg>=0.29.0
kivy==2.3.1
kivymd==1.2.0
gunicorn==21.2.0
py2app==0.28.6 